requirements.txt
# QuizBattle Backend Requirements - Python 3.12 Compatible
# Updated for Render deployment with psycopg3

# Core Flask dependencies - Latest Stable
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1

# Database - Latest Stable
SQLAlchemy==2.0.23

# PostgreSQL driver - psycopg3 with binary and pool support
psycopg[binary,pool]>=3.2

# MongoDB driver - Latest Stable
pymongo==4.6.0

# PDF processing
PyPDF2==3.0.1

# HTTP client for API calls - Latest Stable
requests==2.31.0

# Production WSGI server - Latest Stable
gunicorn==21.2.0

# Environment management - Latest Stable
python-dotenv==1.0.0

# Security and JWT - Latest Stable
PyJWT==2.8.0
cryptography==41.0.7

# Rate limiting and caching
Flask-Limiter==3.5.0
redis==5.0.1

# Monitoring and health checks
flask-healthz==0.0.3

# Data validation
marshmallow==3.20.1
webargs==8.3.0

# Timezone support
pytz==2023.4

# Logging
structlog==23.2.0

# Additional utilities
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
